notification and favorite

+++ controllers/userController.js
@@ -1,4 +1,5 @@
 import { userModel } from "../model/userModel.js";
+import { updatevalidator } from "../validator/user-validator.js";
 
 
 export const verifyEmail = async (req, res, next) => {
@@ -37,11 +38,20 @@ export const getUserProfile = async (req, res, next) => {
 
 
 
-export const updateUserProfile = (req, res, next) => {
+export const updateUserProfile =  async (req, res, next) => {
     try {
-
+        const { id } = req.params
+        const {error,value}= updatevalidator.validate(req.body);
+        if (error) {
+            return res.status(400).json(error);
+        }
+        const user = await userModel.findByIdAndUpdate(id,{user: req.auth.id}, value, { new: true })
+        if (!user) {
+            return res.status(404).json({ message: 'User not found' })
+        }
+        res.status(200).json(user)
     } catch (error) {
-
+        next(error)
     }
 }
 
@@ -50,7 +60,7 @@ export const updateUserProfile = (req, res, next) => {
 export const deleteUserAccount = async (req, res, next) => {
     try {
         const { id } = req.params
-        const user = await userModel.findByIdAndDelete(id)
+        const user = await userModel.findByIdAndDelete(id,{user: req.auth.id})
         if (!user) {
             return res.status(404).json({ message: 'User not found' })
         }
